Separate live vs hermetic CallComposite e2e tests
