+++ packages/react-composites/tests/browser/call/live/Localization.test.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { test } from './fixture';
-import { waitForCallCompositeToLoad, loadCallPage, buildUrl, stableScreenshot } from '../common/utils';
+import { test } from '../fixture';
+import { waitForCallCompositeToLoad, loadCallPage, buildUrl, stableScreenshot } from '../../common/utils';
 import { expect } from '@playwright/test';
 import { v1 as generateGUID } from 'uuid';
